% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{averageDeviationC}
\alias{averageDeviationC}
\title{Deviation}
\usage{
averageDeviationC(indexMap, subspace, alpha, referenceDim, numRuns)
}
\arguments{
\item{indexMap}{Index matrix (see \code{\link{sortedIndexMatrix}}) for the
data objects if ordered by dimension. Each entry of the vector contains
the index to the initial data set. The index is starting with 1.}

\item{subspace}{Numerical vector consisting of the subspace dimensions.
Index for dimensions starts with 1.}

\item{alpha}{Percentage of data objects to remain in the subspace slice
(expected value).}

\item{referenceDim}{Reference attribute upon which the conditions for the
slice will be set.}

\item{numRuns}{number of random subspace slices used to estimate the
deviation.}
}
\value{
double, the deviation for the reference dimension.
}
\description{
Average deviation over random subspace slices for a given reference
attribute.
}
\seealso{
\code{\link{sortedIndexMatrix}}
}

