% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDSS.R
\name{CDSS}
\alias{CDSS}
\title{Cumulated Deviation Subspace Search (CDSS)}
\usage{
CDSS(dm, tau)
}
\arguments{
\item{dm}{deviation matrix, see \code{\link{deviationMatrixC}}}
}
\value{
List of subspaces
}
\description{
Directly mines high quality subspaces from two-dimensional projections only.
}
