% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{combinedScoreAUC}
\alias{combinedScoreAUC}
\title{Area under the ROC curve for combined scoring}
\usage{
combinedScoreAUC(combinationFun, scores, label)
}
\arguments{
\item{combinationFun}{Function to combine LOF scores to final scoring.}

\item{scores}{List of scoring vectors for each subspace.}

\item{label}{Class label.}
}
\value{
Area under the ROC curve
}
\description{
Wrapper to calculate the AUC of a combined scoring.
}
