% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deviationMatrixC}
\alias{deviationMatrixC}
\title{Deviation matrix}
\usage{
deviationMatrixC(indexMap, alpha, numRuns)
}
\arguments{
\item{indexMap}{Index for the data objects if ordered by dimension. Each
entry of the vector contains the index to the initial data set. The index
is starting with 1.}

\item{alpha}{Percentage of data objects to remain in the subspace slice
(expected value).}

\item{numRuns}{number of random subspace slices used to estimate the
deviation.}
}
\value{
NumericMatrix with deviations where the row is the reference
  attribute, column the conditional attribute of the 2-dimensional subspace.
  This matrix is asymmetric in general.
}
\description{
Calculates the deviation for all 2-dimensional projections of the full
space.
}
