% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleAndCombination.R
\name{gaussNorm}
\alias{gaussNorm}
\title{Gauss Normalization}
\usage{
gaussNorm(x)
}
\arguments{
\item{x}{outlier scores to normalize}
}
\value{
normalized outlier scores
}
\description{
Fits a Gauss distribution to normalize score values as suggested by Kriegel et al.
}
\references{
Kriegel, Hans-Peter, Peer Kroger, Erich Schubert, and Arthur
  Zimek. 2011. Interpreting and Unifying Outlier Scores. In Proceedings of
  the 2011 SIAM International Conference on Data Mining, p. 13 - 24.
}

