% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HiCSContrastC}
\alias{HiCSContrastC}
\title{HiCS contrast}
\usage{
HiCSContrastC(indexMap, subspace, alpha, numRuns)
}
\arguments{
\item{indexMap}{Index for the data objects if ordered by dimension. Each
entry of the vector contains the index to the initial data set. The index
is starting with 1.}

\item{subspace}{Numerical vector consisting of the subspace dimensions.
Index for dimensions starts with 1.}

\item{alpha}{Percentage of data objects to remain in the subspace slice
(expected value).}

\item{numRuns}{number of random subspace slices used to estimate the
deviation.}
}
\value{
double, the HiCS contrast for the subspace.
}
\description{
Calculates the HiCS contrast for a given subspace.
}
