% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HiCSMatrixC}
\alias{HiCSMatrixC}
\title{HiCS matrix}
\usage{
HiCSMatrixC(indexMap, alpha, numRuns)
}
\arguments{
\item{indexMap}{Index for the data objects if ordered by dimension. Each
entry of the vector contains the index to the initial data set. The index
is starting with 1.}

\item{alpha}{Percentage of data objects to remain in the subspace slice
(expected value).}

\item{numRuns}{number of random subspace slices used to estimate the
deviation.}
}
\value{
Symmetric NumericMatrix with HiCS contrast that is the average
  deviation when the reference dimension is selected randomly.
}
\description{
Calculates the HiCS contrast for all 2-dimensional projections of the full
space.
}
