% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HiCSSearch}
\alias{HiCSSearch}
\title{HiCS Search}
\usage{
HiCSSearch(indexMap, alpha, numRuns, topkSearch, topkOutput)
}
\arguments{
\item{alpha}{Percentage of data objects to remain in the subspace slice
(expected value).}

\item{numRuns}{number of random subspace slices used to estimate the
subspace contrast.}

\item{topkSearch}{top-k contrast subspaces in the search beam.}

\item{topkOutput}{top-k contrast subspaces in the final output.}

\item{Index}{for the data objects if ordered by dimension. Each
entry of the vector contains the index to the initial data set. The index
is starting with 1.}
}
\value{
List of
\item{subspaces}{List of top-k final subspaces ordered by contrast.}
\item{contrast}{Vector of contrast values of the top-k final subspaces.}
}
\description{
HiCS search following the apriori search scheme. The implementation is
according to the original publication and inspired by the nim-implemtation
\url{https://github.com/bluenote10/HiCS}.
}
\references{
Keller, Fabian, Emmanuel Muller, and Klemens Bohm. 2012. HiCS:
  High Contrast Subspaces for Density-Based Outlier Ranking.
  IEEE International Conference on Data Engineering, 2012
}
\seealso{
\code{\link{HiCSContrastC}}, \code{\link{sortedIndexMatrix}}
}

