% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KSTestC}
\alias{KSTestC}
\title{Kolmogorov-Smirnov test}
\usage{
KSTestC(indexSelection, indexMapReferenceAttribute)
}
\arguments{
\item{indexSelection}{Logical vector that describes the subspace slice on
the data set. A random subspace slice can be genearted by \code{\link{randomSubspaceSliceC}}.}

\item{indexMapReferenceAttribute}{Index vector for the data objects ordered by the
reference dimension. Each entry of the vector contains the index to the
initial data set. The index is starting with 1.}
}
\value{
double Supremum of the absolute distance between the marginal and
  conditional distribution of the reference attribute.
}
\description{
Compares marginal and conditional probability density function for a given
subspace slice.
}
\seealso{
\code{\link{randomSubspaceSliceC}}, \code{\link{sortedIndexMatrix}}
}

