% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierModels.R
\name{LOF}
\alias{LOF}
\title{Local Outlier Factor}
\usage{
LOF(data, k)
}
\arguments{
\item{data}{n x m matrix with n objects and m dimensions}

\item{k}{parameter for k-neighborhood}
}
\value{
vector of local outlier factors
}
\description{
Implements the LOF algorithm from the original publication, simplified by
assuming unique distances
}
\references{
Breunig, Markus M., Hans-Peter Kriegel, Raymond T. Ng, and Jörg
  Sander. 2000. LOF: Identifying Density-Based Local Outliers. SIGMOD Rec. 29
  (2). New York, NY, USA: ACM: 93–104.
}

