% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lofCalculation}
\alias{lofCalculation}
\title{Local outlier factor calculation}
\usage{
lofCalculation(lrd, knnIndex, k)
}
\arguments{
\item{lrd}{vector of local reachability densities}

\item{knnIndex}{n x k Matrix of knn indices}

\item{k}{parameter for k-neighborhood}
}
\value{
vector of local outlier factors
}
\description{
Calculates LOF based on reachability densitiesfor all objects
}

