% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleAndCombination.R
\name{maxCombination}
\alias{maxCombination}
\title{Max Combination Function}
\usage{
maxCombination(x, y)
}
\arguments{
\item{x, y}{list of values}
}
\value{
list containing the pairwise maximum
}
\description{
Wrapper to take maximum of two lists
}

