% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randomSubspaceSliceC}
\alias{randomSubspaceSliceC}
\title{Random selection of a subspace slice}
\usage{
randomSubspaceSliceC(indexMap, subspace, alpha, referenceDim)
}
\arguments{
\item{indexMap}{Index matrix (see \code{\link{sortedIndexMatrix}}) for the data objects if ordered by
dimension. Each entry of the vector contains the index to the
initial data set. The index is starting with 1.}

\item{subspace}{Numerical vector consisting of the subspace dimensions. Index for dimensions starts with 1.}

\item{alpha}{Percentage of data objects to remain in the subspace slice (expected value).}

\item{referenceDim}{Reference attribute upon which the conditions for the slice will be set.}
}
\value{
Logical Vector indicating which data objects remain in the subspace slice.
}
\description{
Creates conditions for all dimensions of a given subspace and a given
reference dimension. Subspace selection is not done on the data set but
rather on the pre calculated index map.
}
\seealso{
\code{\link{sortedIndexMatrix}}
}

