% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reverseNearestNeighborsC}
\alias{reverseNearestNeighborsC}
\title{Reverse Nearest Neighbors}
\usage{
reverseNearestNeighborsC(knn)
}
\arguments{
\item{knn}{n x k Matrix of knn indices}
}
\description{
Extracts the reverse nearest neighbors for a given knn matrix
}

