% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviationFunctions.R
\name{sortedIndex}
\alias{sortedIndex}
\title{Index map for data set}
\usage{
sortedIndex(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
A data.table is returned that consists of indexes for the order of each column. E.g. a data table with the first column (0.3,0.5, 0.1) would get an index of (2,3,1), i.e. the elements sorted by (2,3,1) are in increasing order. The index starts with 1.
}
\description{
The index map allows to create subspace slices without working on the underlying data set but rather on the ordered index.
}

