% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviationFunctions.R
\name{sortedIndexMatrix}
\alias{sortedIndexMatrix}
\title{Index matrix for data set}
\usage{
sortedIndexMatrix(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
\code{\link{sortedIndex}} converted to a matrix
}
\description{
This is a wrapper for \code{\link{sortedIndex}}.
}

