% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleAndCombination.R
\name{sumCombination}
\alias{sumCombination}
\title{Sum Combination Function}
\usage{
sumCombination(x, y)
}
\arguments{
\item{x, y}{list of values}
}
\value{
list containing the pairwise sum
}
\description{
Wrapper to sum of two lists
}
