% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{redundancyAUC}
\alias{redundancyAUC}
\title{Quantified redundancy}
\usage{
redundancyAUC(lofactors, label, combinationFun, scaleFun)
}
\arguments{
\item{lofactors}{list of LOF vectors for each subspace}

\item{label}{class label}

\item{combinationFun}{Function to combine LOF scores to final scoring.}

\item{scaleFun}{Function to scale LOF scores.}
}
\value{
List of \item{initialAUC}{AUC using the full set of subspaces.}
  \item{improvedAUC}{AUC using the remaining set of subspaces.}
  \item{numberInitialSpaces}{Number of initial subspaces.}
  \item{numberRemovedSpaces}{Number of remaining subspaces.}
  \item{removedSpaces}{Index of spaces removed from initial set.}
  \item{remainingSpaces}{Index of spaces remaining from initial set.}
}
\description{
Removes redundant subspaces from the set as long as the final scoring yields
a higher AUC.
}

